// src/types/contents.ts
export type Contents = {
    id: String
    title: String
    subject: String
    keyword: String
    author: String
    language: String
    content_type: String
    organization_name : String
    content_name : String
    content_url : String
    uploaded_by: String
    upload_time: Date
    description : String
    premium : Boolean
    status: Boolean
    deleted: Boolean
    created: Date
  }
  